@@ -3,20 +3,19 @@ using System.Threading;
 using System.Threading.Tasks;
 using Camunda.Worker;
 
-namespace SampleCamundaWorker.Handlers
+namespace SampleCamundaWorker.Handlers;
+
+[HandlerTopics("sayHelloGuest")]
+public class SayHelloGuestHandler : IExternalTaskHandler
 {
-    [HandlerTopics("sayHelloGuest")]
-    public class SayHelloGuestHandler : IExternalTaskHandler
+    public Task<IExecutionResult> HandleAsync(ExternalTask externalTask, CancellationToken cancellationToken)
     {
-        public Task<IExecutionResult> HandleAsync(ExternalTask externalTask, CancellationToken cancellationToken)
+        return Task.FromResult<IExecutionResult>(new CompleteResult
         {
-            return Task.FromResult<IExecutionResult>(new CompleteResult
+            Variables = new Dictionary<string, Variable>
             {
-                Variables = new Dictionary<string, Variable>
-                {
-                    ["MESSAGE"] = Variable.String("Hello, Guest!")
-                }
-            });
-        }
+                ["MESSAGE"] = Variable.String("Hello, Guest!")
+            }
+        });
     }
 }
