@@ -3,30 +3,29 @@ using System.Threading;
 using System.Threading.Tasks;
 using Camunda.Worker;
 
-namespace SampleCamundaWorker.Handlers
+namespace SampleCamundaWorker.Handlers;
+
+[HandlerTopics("sayHello", LockDuration = 10000)]
+[HandlerVariables("USERNAME")]
+public class SayHelloHandler : IExternalTaskHandler
 {
-    [HandlerTopics("sayHello", LockDuration = 10000)]
-    [HandlerVariables("USERNAME")]
-    public class SayHelloHandler : IExternalTaskHandler
+    public async Task<IExecutionResult> HandleAsync(ExternalTask externalTask, CancellationToken cancellationToken)
     {
-        public async Task<IExecutionResult> HandleAsync(ExternalTask externalTask, CancellationToken cancellationToken)
+        if (!externalTask.Variables.TryGetValue("USERNAME", out var usernameVariable))
         {
-            if (!externalTask.Variables.TryGetValue("USERNAME", out var usernameVariable))
-            {
-                return new BpmnErrorResult("NO_USER", "Username not provided");
-            }
-
-            var username = usernameVariable.Value;
-
-            await Task.Delay(1000);
-
-            return new CompleteResult
-            {
-                Variables = new Dictionary<string, Variable>
-                {
-                    ["MESSAGE"] = Variable.String($"Hello, {username}!")
-                }
-            };
+            return new BpmnErrorResult("NO_USER", "Username not provided");
         }
+
+        var username = usernameVariable.Value;
+
+        await Task.Delay(1000);
+
+        return new CompleteResult
+        {
+            Variables = new Dictionary<string, Variable>
+            {
+                ["MESSAGE"] = Variable.String($"Hello, {username}!")
+            }
+        };
     }
 }
