@@ -9,49 +9,48 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using SampleCamundaWorker.Handlers;
 
-namespace SampleCamundaWorker
+namespace SampleCamundaWorker;
+
+public class Startup
 {
-    public class Startup
+    public Startup(IConfiguration configuration)
     {
-        public Startup(IConfiguration configuration)
+        Configuration = configuration;
+    }
+
+    public IConfiguration Configuration { get; }
+
+    public void ConfigureServices(IServiceCollection services)
+    {
+        services.AddExternalTaskClient(client =>
         {
-            Configuration = configuration;
-        }
+            client.BaseAddress = new Uri("http://localhost:8080/engine-rest");
+        });
 
-        public IConfiguration Configuration { get; }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddExternalTaskClient(client =>
-                {
-                    client.BaseAddress = new Uri("http://localhost:8080/engine-rest");
-                });
-
-            services.AddCamundaWorker("sampleWorker")
-                .AddHandler<SayHelloHandler>()
-                .AddHandler<SayHelloGuestHandler>()
-                .ConfigurePipeline(pipeline =>
-                {
-                    pipeline.Use(next => async context =>
-                    {
-                        var logger = context.ServiceProvider.GetRequiredService<ILogger<Startup>>();
-                        logger.LogInformation("Started processing of task {Id}", context.Task.Id);
-                        await next(context);
-                        logger.LogInformation("Finished processing of task {Id}", context.Task.Id);
-                    });
-                });
-
-            services.AddHealthChecks();
-        }
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
+        services.AddCamundaWorker("sampleWorker")
+            .AddHandler<SayHelloHandler>()
+            .AddHandler<SayHelloGuestHandler>()
+            .ConfigurePipeline(pipeline =>
             {
-                app.UseDeveloperExceptionPage();
-            }
+                pipeline.Use(next => async context =>
+                {
+                    var logger = context.ServiceProvider.GetRequiredService<ILogger<Startup>>();
+                    logger.LogInformation("Started processing of task {Id}", context.Task.Id);
+                    await next(context);
+                    logger.LogInformation("Finished processing of task {Id}", context.Task.Id);
+                });
+            });
 
-            app.UseHealthChecks("/health");
+        services.AddHealthChecks();
+    }
+
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
+        {
+            app.UseDeveloperExceptionPage();
         }
+
+        app.UseHealthChecks("/health");
     }
 }
