@@ -4,31 +4,30 @@ using System.Threading.Tasks;
 using Moq;
 using Xunit;
 
-namespace Camunda.Worker.Execution
+namespace Camunda.Worker.Execution;
+
+public class TopicBasedEndpointProviderTest
 {
-    public class TopicBasedEndpointProviderTest
+    [Fact]
+    public void TestGetKnownEndpointDelegate()
     {
-        [Fact]
-        public void TestGetKnownEndpointDelegate()
+        Task FakeHandlerDelegate(IExternalTaskContext context) => Task.CompletedTask;
+
+        var provider = new TopicBasedEndpointProvider(new[]
         {
-            Task FakeHandlerDelegate(IExternalTaskContext context) => Task.CompletedTask;
+            new HandlerDescriptor(FakeHandlerDelegate, new HandlerMetadata(new[] {"topic1"}))
+        });
 
-            var provider = new TopicBasedEndpointProvider(new[]
-            {
-                new HandlerDescriptor(FakeHandlerDelegate, new HandlerMetadata(new[] {"topic1"}))
-            });
+        var handlerDelegate = provider.GetEndpointDelegate(new ExternalTask("test", "test", "topic1"));
+        Assert.NotNull(handlerDelegate);
+    }
 
-            var handlerDelegate = provider.GetEndpointDelegate(new ExternalTask("test", "test", "topic1"));
-            Assert.NotNull(handlerDelegate);
-        }
-
-        [Fact]
-        public void TestGetUnknownEndpointDelegate()
-        {
-            var provider = new TopicBasedEndpointProvider(Enumerable.Empty<HandlerDescriptor>());
-            Assert.Throws<ArgumentException>(
-                () => provider.GetEndpointDelegate(new ExternalTask("test", "test", "topic1"))
-            );
-        }
+    [Fact]
+    public void TestGetUnknownEndpointDelegate()
+    {
+        var provider = new TopicBasedEndpointProvider(Enumerable.Empty<HandlerDescriptor>());
+        Assert.Throws<ArgumentException>(
+            () => provider.GetEndpointDelegate(new ExternalTask("test", "test", "topic1"))
+        );
     }
 }
