@@ -2,23 +2,22 @@ using System.Threading.Tasks;
 using Moq;
 using Xunit;
 
-namespace Camunda.Worker
+namespace Camunda.Worker;
+
+public class NoneResultTest
 {
-    public class NoneResultTest
+    private readonly Mock<IExternalTaskContext> _contextMock = new();
+
+    [Fact]
+    public async Task TestExecuteResultAsync()
     {
-        private readonly Mock<IExternalTaskContext> _contextMock = new();
+        // Arrange
+        var result = new NoneResult();
 
-        [Fact]
-        public async Task TestExecuteResultAsync()
-        {
-            // Arrange
-            var result = new NoneResult();
+        // Act
+        await result.ExecuteResultAsync(_contextMock.Object);
 
-            // Act
-            await result.ExecuteResultAsync(_contextMock.Object);
-
-            // Assert
-            _contextMock.VerifyNoOtherCalls();
-        }
+        // Assert
+        _contextMock.VerifyNoOtherCalls();
     }
 }
